@stephentoub 2018-01-22
Thanks.  As this isn't a public API, what kind of measurable performance impact does this have on the public API that use this?