@gfoidl 2018-01-22
_Linq_ uses the `LargeArrayBuilder<T>` quite a lot. So you mean a benchmark that covers the "public entry" from Linq to see what's the advantage? I make some tests.

I assume that any improvement on LargeArrayBuilder also improves the Linq-methods that use it.